\name{cattle50K}
\docType{data}
\alias{cattle50K}
\title{Genotyped by Bovine50K chip}
\description{
This dataset gives the effects of all SNPs using rrBLUP algorithm for 3 traits, individuals were genotyped by Bovine50K chip.
}
\usage{data(cattle50K)}
\format{A dataframe containing 3 traits' SNPs effects}
\keyword{datasets}