% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_bin}
\alias{.farmcpu_bin}
\title{.farmcpu_bin}
\usage{
.farmcpu_bin(
  Y = NULL,
  GDP = NULL,
  GDP_index = NULL,
  GM = NULL,
  CV = NULL,
  P = NULL,
  method = "EMMA",
  b = c(5e+05, 5e+06, 5e+07),
  s = seq(10, 100, 10),
  theLoop = NULL,
  bound = NULL,
  ncpus = 2,
  verbose = TRUE
)
}
\arguments{
\item{Y}{a n by 2 matrix, the fist column is taxa id and the second is trait}

\item{GDP}{genotype}

\item{GDP_index}{the index of effective genotyped individuals}

\item{GM}{SNP map information, m by 3 matrix, m is marker size, the three columns are SNP_ID, Chr, and Pos}

\item{CV}{covariates, n by c matrix, n is sample size, c is number of covariates}

\item{P}{start p values for all SNPs}

\item{method}{method for selecting the optimal bin: 'static', 'FaST-LMM', 'EMMA', or 'GEMMA'}

\item{b}{bin sizes for all iterations, a vector, the bin size is always from large to small}

\item{s}{number of selected bins in each iteration, a vector}

\item{theLoop}{iteration number}

\item{bound}{maximum number of SNPs selected as pseudo QTNs in each iteration}

\item{ncpus}{number of threads used for parallele computation}

\item{verbose}{whether to print detail.}
}
\value{
Output: seqQTN - an s by 1 vecter for index of QTNs on GM file
}
\description{
Last update: March 28, 2017
Requirement: Y, GDP, and CV must have same taxa order. GDP and GM must have the same order on SNP
Requirement: P and GM are in the same order
Requirement: No missing data
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
\keyword{internal}
