% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_burger}
\alias{.farmcpu_burger}
\title{.farmcpu_burger}
\usage{
.farmcpu_burger(Y = NULL, CV = NULL, GK = NULL, ncpus = 2, method = "FaST-LMM")
}
\arguments{
\item{Y}{phenotype, n by t matrix, n is sample size, t is number of phenotypes}

\item{CV}{covariates, n by c matrix, n is sample size, c is number of covariates}

\item{GK}{Genotype data in numerical format, taxa goes to row and snp go to columns}

\item{ncpus}{number of threads used for parallele computation}

\item{method}{two options for estimating variance components, 'FaST-LMM' or 'EMMA'}
}
\value{
Output: REMLs - maximum log likelihood
Output: vg - genetic variance
Output: ve - residual variance
Output: delta - exp(root)
}
\description{
Last update: Dec 21, 2016
To calculate likelihood, variances and ratio, revised by Xiaolei based on GAPIT.Burger function from GAPIT package
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
\keyword{internal}
