% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_lm}
\alias{.farmcpu_lm}
\title{To quickly sovel LM with one variable substitute multiple times}
\usage{
.farmcpu_lm(
  y,
  w = NULL,
  GDP,
  GDP_index = NULL,
  GDP_mrk_bycol = TRUE,
  maxLine = 5000,
  ncpus = 2,
  npc = 0,
  verbose = TRUE
)
}
\arguments{
\item{y}{one column matrix, dependent variable}

\item{w}{covariates, n by c matrix, n is sample size, c is number of covariates}

\item{GDP}{genotype}

\item{GDP_index}{index of effective genotyped individuals}

\item{GDP_mrk_bycol}{whether the markers are stored by columns in genotype (i.e. M is a n by m matrix)}

\item{maxLine}{the number of markers handled at a time, smaller value would reduce the memory cost}

\item{ncpus}{number of threads used for parallele computation}

\item{npc}{number of covariates without pseudo QTNs}

\item{verbose}{whether to print detail.}
}
\value{
Output: P - p-value of each SNP
Output: betapred - effects of pseudo QTNs
Output: B - effect of each SNP
}
\description{
Start  date: March 1, 2013
Last update: March 6, 2013
Strategy:
1. Separate constant covariates (w) and dynamic coveriates (x)
2. Build non-x related only once
3. Use apply to iterate x
4. Derive dominance indicate d from additive indicate (x) mathmaticaly
5. When d is not estimable, continue to test x
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
\keyword{internal}
