% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_prior}
\alias{.farmcpu_prior}
\title{Set prior on existing p value}
\usage{
.farmcpu_prior(GM, P = NULL, Prior = NULL)
}
\arguments{
\item{GM}{an m by 3  matrix for SNP name, chromosome and BP}

\item{P}{an m by 1 matrix containing probability}

\item{Prior}{an s by 4  matrix for SNP name, chromosome, BP and p-value}
}
\value{
Output: P - updated P value by prior information
}
\description{
Last update: March 10, 2013
Requirement: P and GM are in the same order, Prior is part of GM except P value
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
\keyword{internal}
