% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_remove}
\alias{.farmcpu_remove}
\title{Remove bins that are highly correlated}
\usage{
.farmcpu_remove(
  GDP = NULL,
  GDP_index = NULL,
  GM = NULL,
  seqQTN = NULL,
  seqQTN.p = NULL,
  threshold = 0.99
)
}
\arguments{
\item{GDP}{genotype}

\item{GDP_index}{the index of effective genotyped individuals}

\item{GM}{SNP map information, m by 3 matrix, m is marker size, the three columns are SNP_ID, Chr, and Pos}

\item{seqQTN}{s by 1 vecter for index of QTN on GM}

\item{seqQTN.p}{p value of each @param seqQTN}

\item{threshold}{pearson correlation threshold for remove correlated markers}
}
\value{
Output: bin - n by s0 matrix of genotype
Output: binmap - s0 by 3 matrix for map of bin
Output: seqQTN - s0 by 1 vecter for index of QTN on GM
Relationship: bin=GDP[,c(seqQTN)], binmap=GM[seqQTN,], s0<=s
}
\description{
Last update: March 4, 2013
Requirement: GDP and GM have the same order on SNP
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
\keyword{internal}
