% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_specify}
\alias{.farmcpu_specify}
\title{To get indicator (TURE or FALSE) for GI based on GP}
\usage{
.farmcpu_specify(
  GI = NULL,
  GP = NULL,
  bin.size = 1e+07,
  inclosure.size = NULL,
  MaxBP = 1e+10
)
}
\arguments{
\item{GI}{Data frame with three columns (SNP name, chr and base position)}

\item{GP}{Data frame with seven columns (SNP name, chr and base position, P, MAF, N, effect)}

\item{bin.size}{a value of @param b in 'FarmCPU.bin' function}

\item{inclosure.size}{a value of @param s in 'FarmCPU.bin' function}

\item{MaxBP}{maximum base pairs for each chromosome}
}
\value{
theIndex: a vector indicating if the SNPs in GI belong to QTN or not
}
\description{
Last update: January 26, 2017
Strategy
1.set bins for all snps in GP
2.keep the snp with smallest P value in each bin, record SNP ID
3.Search GI for SNP with SNP ID from above
4.return the position for SNP selected
}
\author{
Zhiwu Zhang
}
