% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{.farmcpu_sub}
\alias{.farmcpu_sub}
\title{.farmcpu_sub}
\usage{
.farmcpu_sub(GM = NULL, GLM = NULL, QTN = NULL, method = "mean")
}
\arguments{
\item{GM}{SNP map information, m by 3 matrix, m is marker size, the three columns are SNP_ID, Chr, and Pos}

\item{GLM}{FarmCPU.GLM object}

\item{QTN}{s by 3  matrix for SNP name, chromosome and BP}

\item{method}{options are "penalty", "reward","mean","median",and "onsite"}
}
\value{
Output: GLM$P - Updated p-values by substitution process
Output: GLM$B - Updated effects by substitution process
}
\description{
Last update: Febuary 26, 2013
Requirement: P has row name of SNP. s<=t. covariates of QTNs are next to SNP
}
\author{
Xiaolei Liu and Zhiwu Zhang
}
