% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FarmCPU.r
\name{FarmCPU.FaSTLMM.LL}
\alias{FarmCPU.FaSTLMM.LL}
\title{FarmCPU.FaSTLMM.LL
Evaluation of the maximum likelihood using FaST-LMM method
 
Last update: January 11, 2017
Requirement: pheno, snp.pool, and X0 must have same taxa order.
Requirement: No missing data}
\usage{
FarmCPU.FaSTLMM.LL(pheno, snp.pool, X0 = NULL, ncpus = 2)
}
\arguments{
\item{pheno}{a two-column phenotype matrix}

\item{snp.pool}{matrix for pseudo QTNs}

\item{X0}{covariates matrix}

\item{ncpus}{number of threads used for parallel computation}
}
\value{
Output: beta - beta effect
Output: delta - delta value
Output: LL - log-likelihood
Output: vg - genetic variance
Output: ve - residual variance
}
\description{
FarmCPU.FaSTLMM.LL
Evaluation of the maximum likelihood using FaST-LMM method
 
Last update: January 11, 2017
Requirement: pheno, snp.pool, and X0 must have same taxa order.
Requirement: No missing data
}
\author{
Xiaolei Liu (modified)
}
