% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.R
\name{MVP.Data.Hapmap2MVP}
\alias{MVP.Data.Hapmap2MVP}
\title{MVP.Data.Hapmap2MVP: To transform Hapmap data to MVP package
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.Hapmap2MVP(hapmap_file, out = "mvp", type.geno = "char",
  verbose = TRUE)
}
\arguments{
\item{hapmap_file}{Genotype in Hapmap format}

\item{out}{the name of output file}

\item{type.geno}{the type of genotype elements}

\item{verbose}{whether to print the reminder}
}
\value{
number of individuals and markers.
Output files:
genotype.desc, genotype.bin: genotype file in bigmemory format
phenotype.phe: ordered phenotype file, same taxa order with genotype file
map.map: SNP information
}
\description{
MVP.Data.Hapmap2MVP: To transform Hapmap data to MVP package
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
hapmapPath <- system.file("extdata", "03_hapmap", "mvp.diploid.hmp.txt", package = "rMVP")
MVP.Data.Hapmap2MVP(hapmapPath, "rMVP.test.hmp")
}
