% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.R
\name{MVP.Data.impute}
\alias{MVP.Data.impute}
\title{MVP.Data.impute: To impute the missing genotype
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.impute(mvp_file, out = "mvp.imp", method = "Major",
  ncpus = NULL)
}
\arguments{
\item{mvp_file}{the name of mvp file}

\item{out}{the name of output file}

\item{method}{'Major', 'Minor', "Middle"}

\item{ncpus}{number of threads for imputing}
}
\value{
NULL
Output files:
imputed genotype file
}
\description{
MVP.Data.impute: To impute the missing genotype
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
mvpPath <- system.file("extdata", "mvp.geno.desc", package = "rMVP")
MVP.Data.impute(mvpPath)
}
