% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.R
\name{MVP.Data.Map}
\alias{MVP.Data.Map}
\title{MVP.Data.Map: To check map file
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.Map(map_file, out = "mvp", cols = c(1, 2, 3), header = TRUE,
  sep = "\\t")
}
\arguments{
\item{map_file}{the name of map file}

\item{out}{the name of output file}

\item{cols}{selected columns}

\item{header}{whether the file contains header}

\item{sep}{seperator of the file}
}
\value{
NULL
}
\description{
MVP.Data.Map: To check map file
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
mapPath <- system.file("extdata", "mvp.map", package = "rMVP")
MVP.Data.Map(mvp.map)
}
