% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.r
\name{MVP.Data.MVP2Bfile}
\alias{MVP.Data.MVP2Bfile}
\title{MVP.Data.MVP2Bfile: To transform MVP data to binary format
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.MVP2Bfile(bigmat, map, pheno = NULL, out = "mvp.plink",
  verbose = TRUE)
}
\arguments{
\item{bigmat}{Genotype in bigmatrix format (0,1,2)}

\item{map}{the map file}

\item{pheno}{the phenotype file}

\item{out}{the name of output file}

\item{verbose}{whether to print the reminder}
}
\value{
NULL
Output files:
.bed, .bim, .fam
}
\description{
MVP.Data.MVP2Bfile: To transform MVP data to binary format
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
bigmat <- as.big.matrix(matrix(1:6, 3, 2))
map <- matrix(c("rs1", "rs2", "rs3", 1, 1, 1, 10, 20, 30), 3, 3)

MVP.Data.MVP2Bfile(bigmat, map, out=tempfile("outfile"))

}
