% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.r
\name{MVP.Data.Pheno}
\alias{MVP.Data.Pheno}
\title{MVP.Data.Pheno: To clean up phenotype file
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.Pheno(pheno_file, out = "mvp", cols = NULL, header = TRUE,
  sep = "\\t", missing = c(NA, "NA", "-9", 9999), verbose = TRUE)
}
\arguments{
\item{pheno_file}{the name of phenotype file}

\item{out}{the name of output file}

\item{cols}{selected columns}

\item{header}{whether the file contains header}

\item{sep}{seperator of the file}

\item{missing}{the missing value}

\item{verbose}{whether to print detail.}
}
\value{
NULL
Output files:
cleaned phenotype file
}
\description{
MVP.Data.Pheno: To clean up phenotype file
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
phePath <- system.file("extdata", "07_other", "mvp.phe", package = "rMVP")

MVP.Data.Pheno(phePath, out=tempfile("outfile"))

}
