% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.R
\name{MVP.Data.QC}
\alias{MVP.Data.QC}
\title{MVP.Data.QC: quality control of genotype
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.QC(mvp_prefix, out = NULL, geno = 0.1, mind = 0.1,
  maf = 0.05, hwe = NULL, ncpus = NULL)
}
\arguments{
\item{mvp_prefix}{the prefix of genotype file}

\item{out}{the prefix of output file}

\item{geno}{the threshold of calling rate of markers}

\item{mind}{the threshold of calling rate of individuals}

\item{maf}{the threshold of minor allel frequency}

\item{hwe}{the threshold of hwe}

\item{ncpus}{the number of threads for quality control}
}
\value{
NULL
Output files:
cleaned genotype file
}
\description{
MVP.Data.QC: quality control of genotype
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
geno <- file.path(system.file(), "06_mvp-impute", "mvp.imp")
MVP.Data.QC(geno, out="myKin")
}
