% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.R
\name{MVP.Data}
\alias{MVP.Data}
\title{MVP.Data: To prepare data for MVP package
Author: Xiaolei Liu, Lilin Yin and Haohao Zhang
Build date: Aug 30, 2016
Last update: Sep 12, 2018}
\usage{
MVP.Data(fileMVP = NULL, fileVCF = NULL, fileHMP = NULL,
  fileBed = NULL, fileNum = NULL, fileMap = NULL, filePhe = NULL,
  fileInd = NULL, fileKin = TRUE, filePC = TRUE, out = "mvp",
  sep.num = "\\t", auto_transpose = TRUE, sep.map = "\\t",
  sep.phe = "\\t", sep.kin = "\\t", sep.pc = "\\t",
  type.geno = "char", pheno_cols = NULL, SNP.impute = "Major",
  maxLine = 10000, priority = "speed", perc = 1, pcs.keep = 5, ...)
}
\arguments{
\item{fileVCF}{Genotype in VCF format}

\item{fileHMP}{Genotype in hapmap format}

\item{fileBed}{Genotype in PLINK binary format}

\item{fileNum}{Genotype in numeric format; pure 0, 1, 2 matrix; m * n, m is marker size, n is sample size}

\item{fileMap}{SNP map information, there are three columns, including SNP_ID, Chromosome, and Position}

\item{filePhe}{Phenotype, the first column is taxa name, the subsequent columns are traits}

\item{fileKin}{Kinship that represents relationship among individuals, n * n matrix, n is sample size}

\item{filePC}{Principal components, n*npc, n is sample size, npc is number of top columns of principal components}

\item{out}{A marker on output file name}

\item{sep.num}{seperator for numeric file.}

\item{sep.map}{seperator for map file.}

\item{sep.phe}{seperator for phenotype file.}

\item{sep.kin}{seperator for Kinship file.}

\item{sep.pc}{seperator for PC file.}

\item{type.geno}{type parameter in bigmemory, genotype data}

\item{SNP.impute}{"Left", "Middle", "Right"}

\item{maxLine}{number of SNPs, only used for saving memory when calculate kinship matrix}

\item{priority}{"speed" or "memory"}

\item{perc}{Percentage of markers used to calculate PCA}

\item{pcs.keep}{how many PCs to keep

Output files:
genotype.desc, genotype.bin: genotype file in bigmemory format
phenotype.phe: ordered phenotype file, same taxa order with genotype file
map.map: SNP information
k.desc, k.bin: Kinship matrix in bigmemory format
pc.desc, pc.bin: PC matrix in bigmemory format
Requirement: fileHMP, fileBed, and fileNum can not input at the same time}

\item{sep.vcf}{seperator for vcf file.}

\item{type.kin}{type parameter in bigmemory, Kinship}

\item{type.pc}{type parameter in bigmemory, PC}

\item{type.map}{type parameter in bigmemory, PC}

\item{SNP.effect}{"Add" or "Dom"}

\item{maxRecord}{maximum number for markers}

\item{maxInd}{maximum number for individuals}
}
\description{
MVP.Data: To prepare data for MVP package
Author: Xiaolei Liu, Lilin Yin and Haohao Zhang
Build date: Aug 30, 2016
Last update: Sep 12, 2018
}
\examples{
MVP.Data
}
