% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.GEMMA.Vg.Ve.R
\name{MVP.GEMMA.Vg.Ve}
\alias{MVP.GEMMA.Vg.Ve}
\title{To estimate variance component using HE regression}
\usage{
MVP.GEMMA.Vg.Ve(y, X, K, rtol = 1e-06, atol = 1e-08, ctol = 1e-08)
}
\arguments{
\item{y}{phenotype}

\item{X}{genotype}

\item{K}{kinship matrix}

\item{rtol}{parameters for HE regression, no changes is recommended}

\item{atol}{parameters for HE regression, no changes is recommended}

\item{ctol}{parameters for HE regression, no changes is recommended}
}
\value{
vg, ve, and delta
}
\description{
Build date: Feb 2, 2017
Last update: Feb 2, 2017
}
\examples{
phePath <- system.file("extdata", "mvp.phe", package = "rMVP")
phenotype <- read.table(phePath, header=TRUE)
print(dim(phenotype))
genoPath <- system.file("extdata", "mvp.geno.desc", package = "rMVP")
genotype <- attach.big.matrix(genoPath)
print(dim(genotype))
K <- MVP.K.VanRaden(genotype)
vc <- MVP.GEMMA.Vg.Ve(y=phenotype[,2], X=matrix(1, nrow(phenotype)), K=K)
print(vc)
}
\author{
Translated from C++(GEMMA, Xiang Zhou) to R by: Haohao Zhang
}
