% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Utility.r
\name{paste_label}
\alias{paste_label}
\title{Paste label to a line}
\usage{
paste_label(line, label, side = "right", margin = 2)
}
\arguments{
\item{line}{long text}

\item{label}{short label}

\item{side}{"right" or "left"}

\item{margin}{default 2}
}
\description{
Build date: Oct 22, 2018
Last update: Oct 22, 2018
}
\author{
Haohao Zhang
}
\keyword{internal}
