\name{pig60K}
\docType{data}
\alias{pig60K}
\title{Genotyped by pig 60k chip}
\description{
This dataset gives the results of Genome-wide association study of 3 traits, individuals were genotyped by pig 60K chip.
}
\usage{data(pig60K)}
\format{A dataframe containing 3 traits' Pvalue}
\keyword{datasets}