% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{is.big.matrix}
\alias{as.big.matrix}
\alias{attach.big.matrix}
\alias{deepcopy}
\alias{new}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bigmemory}{\code{\link[bigmemory:as.big.matrix]{as.big.matrix()}}, \code{\link[bigmemory:attach.big.matrix]{attach.big.matrix()}}, \code{\link[bigmemory:deepcopy]{deepcopy()}}, \code{\link[bigmemory:is.big.matrix]{is.big.matrix()}}}

  \item{methods}{\code{\link[methods:new]{new()}}}
}}

