% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Data.r
\name{MVP.Data.Map}
\alias{MVP.Data.Map}
\title{MVP.Data.Map: To check map file
Author: Haohao Zhang
Build date: Sep 12, 2018}
\usage{
MVP.Data.Map(map, out = "mvp", cols = c(1, 2, 3), header = TRUE,
  sep = "\\t", verbose = TRUE)
}
\arguments{
\item{map}{the name of map file or map object(data.frame or matrix)}

\item{out}{the name of output file}

\item{cols}{selected columns}

\item{header}{whether the file contains header}

\item{sep}{seperator of the file}

\item{verbose}{whether to print detail.}
}
\value{
Output file:
<out>.map
}
\description{
MVP.Data.Map: To check map file
Author: Haohao Zhang
Build date: Sep 12, 2018
}
\examples{
mapPath <- system.file("extdata", "07_other", "mvp.map", package = "rMVP")

MVP.Data.Map(mapPath, file.path(tempdir(), "rMVP.test"))

}
