\name{MVP.EMMA.Vg.Ve}
\alias{MVP.EMMA.Vg.Ve}
\title{Estimate variance components using EMMA}
\description{
Estimate variance components using EMMA
}
\usage{
MVP.EMMA.Vg.Ve(y, X, K, ngrids=100, llim=-10, ulim=10, esp=1e-10)
}
\arguments{
	\item{y}{phenotype}
	\item{X}{genotype}
	\item{K}{kinship matrix}
	\item{rtol}{parameters for HE regression, no changes is recommended}
	\item{atol}{parameters for HE regression, no changes is recommended}
	\item{ctol}{parameters for HE regression, no changes is recommended}
	\item{root}{parameters for HE regression, no changes is recommended}
}
\author{
Lilin Yin, Zhiwu Zhang, Xinyun Li, Shuhong Zhao, and Xiaolei Liu \cr
Maintainer: Xiaolei Liu <\bold{xiaoleiliu@mail.hzau.edu.cn}>
}
