% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.FaSTLMM.LL.r
\name{MVP.FaSTLMM.LL}
\alias{MVP.FaSTLMM.LL}
\title{Evaluation of the maximum likelihood using FaST-LMM method}
\usage{
MVP.FaSTLMM.LL(pheno, snp.pool, X0 = NULL, ncpus = 2)
}
\arguments{
\item{pheno}{a two-column phenotype matrix}

\item{snp.pool}{matrix for pseudo QTNs}

\item{X0}{covariates matrix}

\item{ncpus}{number of threads used for parallel computation}
}
\value{
Output: beta - beta effect
Output: delta - delta value
Output: LL - log-likelihood
Output: vg - genetic variance
Output: ve - residual variance
}
\description{
Last update: January 11, 2017
}
\author{
Xiaolei Liu (modified)
}
