\name{MVP.FaSTLMM.LL}
\alias{MVP.FaSTLMM.LL}
\title{Evaluation of the maximum likelihood using FaST-LMM method}
\description{
Evaluation of the maximum likelihood using FaST-LMM method
}
\usage{
MVP.FaSTLMM.LL(pheno, snp.pool,X0=NULL,ncpus=2)
}
\arguments{
	\item{pheno}{phenotype}
	\item{snp.pool}{SNP matrix}
	\item{X0}{covariates matrix}
	\item{ncpus}{number of CPUs used for parallel computation}
}
\author{
Lilin Yin, Zhiwu Zhang, Xinyun Li, Shuhong Zhao, and Xiaolei Liu \cr
Maintainer: Xiaolei Liu <\bold{xiaoleiliu@mail.hzau.edu.cn}>
}
