\name{MVP.GEMMA.Vg.Ve}
\alias{MVP.GEMMA.Vg.Ve}
\title{To estimate variance component using HE regression}
\description{
To estimate variance component using HE regression
}
\usage{
MVP.GEMMA.Vg.Ve(y, X, K, rtol=1e-6, atol=1e-8, ctol=1e-8, root=FALSE)
}
\arguments{
	\item{y}{phenotype}
	\item{X}{genotype}
	\item{K}{kinship matrix}
	\item{rtol}{parameters for HE regression, no changes is recommended}
	\item{atol}{parameters for HE regression, no changes is recommended}
	\item{ctol}{parameters for HE regression, no changes is recommended}
	\item{root}{parameters for HE regression, no changes is recommended}
}
\author{
Lilin Yin, Zhiwu Zhang, Xinyun Li, Shuhong Zhao, and Xiaolei Liu \cr
Maintainer: Xiaolei Liu <\bold{xiaoleiliu@mail.hzau.edu.cn}>
}
