% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.K.VanRaden.R
\name{MVP.K.VanRaden}
\alias{MVP.K.VanRaden}
\title{Calculate Kinship matrix by VanRaden method}
\usage{
MVP.K.VanRaden(M, weight = NULL, priority = c("speed", "memory"),
  memo = NULL, SUM = NULL, maxLine = 1000)
}
\arguments{
\item{M}{Genotype, m * n, m is marker size, n is population size}

\item{weight}{vector, the weights for makers}

\item{priority}{speed or memory}

\item{memo}{add a character to the name of temporary files}

\item{SUM}{the scaled value to kinship matrix}

\item{maxLine}{when the priority is 'memory', users can change this parameter to limit the memory}
}
\value{
K, n * n matrix
}
\description{
Build date: Dec 12, 2016
Last update: Dec 12, 2016
}
\examples{
genoPath <- system.file("extdata", "mvp.geno.desc", package = "rMVP")
genotype <- attach.big.matrix(genoPath)
print(dim(genotype))
K <- MVP.K.VanRaden(genotype)
}
