% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Report.r
\name{MVP.Report}
\alias{MVP.Report}
\title{MVP.Report}
\usage{
MVP.Report(MVP, col = c("#4197d8", "#f8c120", "#413496", "#495226",
  "#d60b6f", "#e66519", "#d581b7", "#83d3ad", "#7c162c", "#26755d"),
  bin.size = 1e+06, bin.range = NULL, pch = 19, band = 1,
  H = 1.5, ylim = NULL, cex.axis = 1, lwd.axis = 1.5,
  cex.lab = 1.5, plot.type = "b", multracks = FALSE, cex = c(0.5,
  1, 1), r = 0.3, xlab = "Chromosome",
  ylab = expression(-log[10](italic(p))), xaxs = "i", yaxs = "r",
  outward = FALSE, threshold = NULL, threshold.col = "red",
  threshold.lwd = 1, threshold.lty = 2, amplify = FALSE,
  signal.cex = 1.5, signal.pch = 19, signal.col = "red",
  signal.line = 1, highlight = NULL, highlight.cex = 1.5,
  highlight.pch = 19, highlight.col = "green", chr.labels = NULL,
  chr.den.col = "black", cir.band = 1, cir.chr = TRUE,
  cir.chr.h = 1.5, cir.legend = TRUE, cir.legend.cex = 0.6,
  cir.legend.col = "black", LOG10 = TRUE, box = FALSE,
  conf.int = TRUE, file.output = TRUE, outpath = getwd(),
  file.type = "jpg", dpi = 300, height = NULL, width = NULL,
  memo = "", verbose = TRUE)
}
\arguments{
\item{MVP}{a dataframe or list, at least four columns. The first column is the name of SNP, the second column is the chromosome of SNP, the third column is the position of SNP, and the remaining columns are the P-value of each trait(Note:each trait a column).}

\item{col}{a vector or a matrix, if "col" is a vector, each circle use the same colors, it means that the same chromosome is drewed in the same color, the colors are not fixed, one, two, three or more colors can be used, if the length of the "col" is shorter than the length the chromosome, then colors will be applied circularly. If "col" is a matrix, the row is the number of circles(traits), the columns are the colors that users want to use for different circles, each circle can be plotted in different number of colors, the missing value can be replaced by NA. For example: col=matrix(c("grey30","grey60",NA,"red","blue","green","orange",NA,NA),3,3,byrow=T).}

\item{bin.size}{the size of bin for SNP_density plot.}

\item{bin.range}{a vector, c(min, max). The min/max value of legend of SNP_density plot, the bin whose SNP number is smaller/bigger than 'bin.range' will be use the same color.}

\item{pch}{a number, the type for the points or for traits of multi-traits Manhattan plot, is the same with "pch" in <plot>.}

\item{band}{a number, the space between chromosomes, the default is 1(if the band equals to 0, then there would be no space between chromosomes).}

\item{H}{a number, the height for each circle, each circle represents a trait, the default is 1.}

\item{ylim}{a vector, the range of Y-axis when plotting the two type of Manhattan plots, is the same with "ylim" in <plot>.}

\item{cex.axis}{a number, controls the size of ticks' numbers of X/Y-axis and the size of labels of circle plot.}

\item{lwd.axis}{a number, controls the width of X/Y-axis lines.}

\item{cex.lab}{a number, controls the size of labels of X/Y-axis.}

\item{plot.type}{a character or vector, only "d", "c", "m", "q" or "b" can be used. if plot.type="d", SNP density will be plotted; if plot.type="c", only circle-Manhattan plot will be plotted; if plot.type="m",only Manhattan plot will be plotted; if plot.type="q",only Q-Q plot will be plotted;if plot.type="b", both circle-Manhattan, Manhattan and Q-Q plots will be plotted; if plot.type=c("m","q"), Both Manhattan and Q-Q plots will be plotted.}

\item{multracks}{a logical,if multracks=FALSE, all Manhattan plots will be drew in separated files, if it is TRUE, all Manhattan plots will be plotted in only one file.}

\item{cex}{a number or a vector, the size for the points, is the same with "size" in <plot>, and if it is a vector, the first number controls the size of points in circle plot(the default is 0.5), the second number controls the size of points in Manhattan plot(the default is 1), the third number controls the size of points in Q-Q plot(the default is 1)}

\item{r}{a number, the radius for the circle(the inside radius), the default is 1.}

\item{xlab}{a character, the labels for x axis.}

\item{ylab}{a character, the labels for y axis.}

\item{xaxs}{a character, The style of axis interval calculation to be used for the x-axis. Possible values are "r", "i", "e", "s", "d". The styles are generally controlled by the range of data or xlim, if given.}

\item{yaxs}{a character, The style of axis interval calculation to be used for the y-axis. See xaxs above..}

\item{outward}{logical, if outward=TRUE,then all points will be plotted from inside to outside for circular Manhattan plot.}

\item{threshold}{a number or vector, the significant threshold. For example, Bonfferoni adjustment method: threshold=0.01/nrow(Pmap). More than one significant line can be added on the plots, if threshold=0 or NULL, then the threshold line will not be added.}

\item{threshold.col}{a character or vector, the colour for the line of threshold levels.}

\item{threshold.lwd}{a number or vector, the width for the line of threshold levels.}

\item{threshold.lty}{a number or vector, the type for the line of threshold levels.}

\item{amplify}{logical, CMplot can amplify the significant points, if amplify=T, then the points bigger than the minimal significant level will be amplified, the default: amplify=TRUE.}

\item{signal.cex}{a number, if amplify=TRUE, users can set the size of significant points.}

\item{signal.pch}{a number, if amplify=TRUE, users can set the shape of significant points.}

\item{signal.col}{a character, if amplify=TRUE, users can set the colour of significant points, if signal.col=NULL, then the colors of significant points will not be changed.}

\item{signal.line}{a number, the width of the lines of significant SNPs cross the circle.}

\item{highlight}{a vector, names of SNPs which need to be highlighted.}

\item{highlight.cex}{a number or vector, the size of points for SNPs which need to be highlighted.}

\item{highlight.pch}{a number or vector, the pch of points for SNPs which need to be highlighted.}

\item{highlight.col}{a number or vector, the col of points for SNPs which need to be highlighted.}

\item{chr.labels}{a vector, the labels for the chromosomes of density plot and circle-Manhattan plot.}

\item{chr.den.col}{a character or vector or NULL, the colour for the SNP density. If the length of parameter 'chr.den.col' is bigger than 1, SNP density that counts the number of SNP within given size('bin.size') will be plotted around the circle. If chr.den.col=NULL, the density bar will not be attached on the bottom of manhattan plot.}

\item{cir.band}{a number, the space between circles, the default is 1.}

\item{cir.chr}{logical, a boundary that represents chromosomes will be plotted on the periphery of a circle, the default is TRUE.}

\item{cir.chr.h}{a number, the width for the boundary, if cir.chr=FALSE, then this parameter will be useless.}

\item{cir.legend}{logical, whether to add the legend of each circle.}

\item{cir.legend.cex}{a number, the size of the number of legend.}

\item{cir.legend.col}{a character, the color of the axis of legend.}

\item{LOG10}{logical, whether to change the p-value into log10(p-value).}

\item{box}{logical, this function draws a box around the current Manhattan plot.}

\item{conf.int}{logical, whether to plot confidence interval on QQ-plot.}

\item{file.output}{a logical, users can choose whether to output the plot results.}

\item{outpath}{Only when file.output = TRUE, determines the path of the output file}

\item{file.type}{a character, users can choose the different output formats of plot, so for, "jpg", "pdf", "tiff" can be selected by users.}

\item{dpi}{a number, the picture resolution for .jpg and .tiff files. The default is 300.}

\item{height}{the height of output files.}

\item{width}{the width of output files.}

\item{memo}{add a character to the output file name.}

\item{verbose}{whether to print the reminder.}
}
\value{
Output files
}
\description{
MVP.Report
}
\examples{
data(pig60K, package = "rMVP")

MVP.Report(pig60K[,c(1:3, 5)], plot.type="m",
    threshold=0.05/nrow(pig60K), file.output=FALSE)

}
