% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Utility.r
\name{MVP.Version}
\alias{MVP.Version}
\title{Print MVP Banner}
\usage{
MVP.Version(width = 60, verbose = TRUE)
}
\arguments{
\item{width}{the width of the message}

\item{verbose}{whether to print detail.}
}
\value{
version number.
}
\description{
Build date: Aug 30, 2017
Last update: Dec 12, 2018
}
\examples{
MVP.Version()
}
\author{
Lilin Yin, Haohao Zhang, and Xiaolei Liu
}
