% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVP.Utility.r
\name{print_info}
\alias{print_info}
\title{Print R Package information, include title, short_title, logo, version, authors, contact}
\usage{
print_info(
  welcome = NULL,
  title = NULL,
  short_title = NULL,
  logo = NULL,
  version = NULL,
  authors = NULL,
  contact = NULL,
  linechar = "=",
  width = NULL,
  verbose = TRUE
)
}
\arguments{
\item{welcome}{welcome text, for example: "Welcom to <Packagename>"}

\item{title}{long text to introduct package}

\item{short_title}{short label, top-left of logo}

\item{logo}{logo}

\item{version}{short label, bottom-right of logo}

\item{contact}{email or website}

\item{width}{banner width}
}
\description{
Build date: Oct 22, 2018
Last update: Oct 22, 2018
}
\author{
Haohao Zhang
}
\keyword{internal}
